Create of-example.yml
Using the Orb Development Kit, adding a new usage example is as simple as creating a new file name-of-example.yml within the orb project’s src/examples directory. Usage examples are not for use in project configuration directly, but are a type of orb metadata to share how a user could best make use of the orb in their configuration and are displayed, for reference, in the Orb Registry.